requirements.txt
requests
feedparser
pandas
python-dateutil
pytz
beautifulsoup4
plotly
streamlit
supabase
python-dotenv
